refactor(generator): relocate pagination code to new file
