+++ generator/google_cloud_cpp_generator.bzl
@@ -46,6 +46,7 @@ google_cloud_cpp_generator_hdrs = [
     "internal/mock_connection_generator.h",
     "internal/option_defaults_generator.h",
     "internal/options_generator.h",
+    "internal/pagination.h",
     "internal/predicate_utils.h",
     "internal/printer.h",
     "internal/proto_definition_location.h",
@@ -94,6 +95,7 @@ google_cloud_cpp_generator_srcs = [
     "internal/mock_connection_generator.cc",
     "internal/option_defaults_generator.cc",
     "internal/options_generator.cc",
+    "internal/pagination.cc",
     "internal/predicate_utils.cc",
     "internal/resolve_comment_references.cc",
     "internal/resolve_method_return.cc",
