+++ generator/internal/descriptor_utils.cc
@@ -35,6 +35,7 @@
 #include "generator/internal/mock_connection_generator.h"
 #include "generator/internal/option_defaults_generator.h"
 #include "generator/internal/options_generator.h"
+#include "generator/internal/pagination.h"
 #include "generator/internal/predicate_utils.h"
 #include "generator/internal/resolve_method_return.h"
 #include "generator/internal/retry_traits_generator.h"
