+++ generator/internal/resolve_method_return.cc
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include "generator/internal/resolve_method_return.h"
+#include "generator/internal/pagination.h"
 #include "generator/internal/predicate_utils.h"
 #include <google/longrunning/operations.pb.h>
 #include <google/protobuf/descriptor.h>
