+++ generator/internal/tracing_connection_generator.cc
@@ -14,6 +14,7 @@
 
 #include "generator/internal/tracing_connection_generator.h"
 #include "generator/internal/codegen_utils.h"
+#include "generator/internal/pagination.h"
 #include "generator/internal/predicate_utils.h"
 #include "generator/internal/printer.h"
 #include "google/cloud/internal/absl_str_cat_quiet.h"
